[encryption-service] Handle non-finished encryption better.
If home encryption is stopped before fstab and crypttab are written then
marker file should not be removed, so that encrypting can be tried
again.

If nemo's home can not be copied to /home partition change the home
directory back to /home/nemo just in case /home was unmounted but the
content was not destroyed.

In preparation script, chmod the copied /etc/skel because othewise
nemo's home may have bad permissions that prevent ssh from working
properly. Also don't fail preparation service if /home can not be
unmounted if it doesn't exist.

Add After=home-encryption-preparation.service to enryption-service's
service file to make sure they are not started or stopped in wrong
order. This is just to make this more robust, in practice they should be
in correct order already.

+++ encryption-service/preparation/home-encryption-finish.sh
@@ -1,6 +1,15 @@
 #!/bin/sh
+
+# If encryption finished, remove marker file, otherwise exit
+[ -s /etc/crypttab ] || exit 4
 rm -f /var/lib/sailfish-device-encryption/encrypt-home
-mount | grep -q " on /home type" || exit 4
-usermod --move-home --home /home/nemo nemo
-mv --target-directory=/home/ /tmp/home/.[!.]* /tmp/home/*
-add-oneshot --user --late preload-ambience
+
+# Move /home back to /home partition if it is mounted
+if $(mount | grep -q " on /home type"); then
+    usermod --move-home --home /home/nemo nemo
+    mv --target-directory=/home/ /tmp/home/.[!.]* /tmp/home/*
+    add-oneshot --user --late preload-ambience
+else
+    usermod --home /home/nemo nemo
+    exit 5
+fi
